pyproject.toml
[tool.poetry]
name = "savannah-test"
version = "0.1.0"
description = ""
authors = ["mogendi <mogendi.mongare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
celery = "^5.3.6"
django-rest-framework = "^0.1.0"
autoflake = "^2.2.1"
django-stubs = "^4.2.7"
django-extensions = "^3.2.3"
factory-boy = "^3.3.0"
coverage = "^7.4.1"
django-auth-oidc = "^0.6.0"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
